Changed title of Ivan's talk
See Ivan's comment:

Although I do not see any problem with the current title you have, only thing that eventually might be better is to define it in a bit more general way as:

  Monad Transformers and the standard transformer libraries (operational/transformers/mtl)

I think the talk will anyway have to introduce concept of monad transformers before getting into the libraries. As well, beginners will be more attracted to the more general title and will certainly gain more by covering the conceptual part.

Still main target of the talk would be to introduce and clarify the best practices and the state of the libraries.